Move Eth-flow typings to file
